The processor module was getting out of hand, so re-org the pipeline processor code into a new module.
