+++ docs/source/pypond.processor.rst
@@ -1,0 +1,94 @@
+pypond.processor package
+========================
+
+Submodules
+----------
+
+pypond.processor.aggregator module
+----------------------------------
+
+.. automodule:: pypond.processor.aggregator
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pypond.processor.base module
+----------------------------
+
+.. automodule:: pypond.processor.base
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pypond.processor.collapser module
+---------------------------------
+
+.. automodule:: pypond.processor.collapser
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pypond.processor.converter module
+---------------------------------
+
+.. automodule:: pypond.processor.converter
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pypond.processor.filler module
+------------------------------
+
+.. automodule:: pypond.processor.filler
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pypond.processor.filter module
+------------------------------
+
+.. automodule:: pypond.processor.filter
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pypond.processor.mapper module
+------------------------------
+
+.. automodule:: pypond.processor.mapper
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pypond.processor.offset module
+------------------------------
+
+.. automodule:: pypond.processor.offset
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pypond.processor.selector module
+--------------------------------
+
+.. automodule:: pypond.processor.selector
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pypond.processor.taker module
+-----------------------------
+
+.. automodule:: pypond.processor.taker
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: pypond.processor
+    :members:
+    :undoc-members:
+    :show-inheritance:
