+++ pypond/processor/offset.py
@@ -1,0 +1,86 @@
+"""
+Simple processor to change the event values by a certain offset.
+
+Primarily for testing.
+"""
+
+from .base import Processor
+from ..event import Event
+from ..exceptions import ProcessorException
+from ..util import is_pipeline, Options
+
+
+class Offset(Processor):
+    """
+    A simple processor used by the testing code to verify Pipeline behavior.
+    """
+
+    def __init__(self, arg1, options=Options()):
+        """A simple processor used by the testing code to verify Pipeline behavior.
+
+        Parameters
+        ----------
+        arg1 : Offset or Pipeline
+            Pipeline or copy constructor
+        options : Options, optional
+            Pipeline Options object.
+
+        Raises
+        ------
+        ProcessorException
+            Raised on bad arg types.
+        """
+        super(Offset, self).__init__(arg1, options)
+
+        self._log('Offset.init', 'uid: {0}'.format(self._id))
+
+        self._by = None
+        self._field_spec = None
+
+        if isinstance(arg1, Offset):
+            # Copy constructor
+            # pylint: disable=protected-access
+            self._by = arg1._by
+            self._field_spec = arg1._field_spec
+        elif is_pipeline(arg1):
+            self._by = options.by
+            self._field_spec = options.field_spec
+        else:
+            msg = 'Unknown arg to Offset constructor: {a}'.format(a=arg1)
+            raise ProcessorException(msg)
+
+    def clone(self):
+        """Clone this Offset processor.
+
+        Returns
+        -------
+        Offset
+            Cloned offset object.
+        """
+        return Offset(self)
+
+    def add_event(self, event):
+        """
+        Output an even that is offset by a certain value.
+
+        Parameters
+        ----------
+        event : Event, IndexedEvent, TimerangeEvent
+            Any of the three event variants.
+        """
+
+        self._log('Offset.add_event', event)
+
+        if self.has_observers():
+            selected = Event.selector(event, self._field_spec)
+            data = dict()
+
+            for k, v in list(selected.data().items()):
+                offset_value = v + self._by
+                data[k] = offset_value
+
+            output_event = event.set_data(data)
+
+            self._log('Offset.add_event', 'emitting: {0}'.format(output_event))
+
+            self.emit(output_event)
